requirements.txt
kconfiglib==14.1.0
requests==2.26.0
urllib3==2.0.2
chardet==4.0.0
pyelftools==0.31